requirements.txt
google-generativeai
streamlit
skyfield
numpy
